"(front end) render dashboard component when user logs in"
