+++ .env
@@ -2,10 +2,11 @@ MY_DB=Paws
 MY_HOST=paws.cykfkgwxg7am.us-east-2.rds.amazonaws.com
 MY_UN=root
 MY_PW=orangetrainfalcon
-APIKEY=AIzaSyANFsNxSnbvx8JYla2srLVtIS2tIOGSPMA,
-AUTHDOMAIN=found-dog.firebaseapp.com,
-DATABSEURL=https://found-dog.firebaseio.com,
-PROJECTID=found-dog,
-STORAGEBUCKET=found-dog.appspot.com,
-MESSAGINGSENDERID=190867854845,
-APPID=:190867854845:web:c995c8b214fa6fe4
+API_KEY=AIzaSyANFsNxSnbvx8JYla2srLVtIS2tIOGSPMA,
+AUTH_DOMAIN=found-dog.firebaseapp.com,
+DATABSE_URL=https://found-dog.firebaseio.com,
+PROJECT_ID=found-dog,
+STORAGE_BUCKET=found-dog.appspot.com,
+MESSAGING_SENDERID=190867854845,
+APP_ID=:190867854845:web:c995c8b214fa6fe4
+PORT=5432
