+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
 .env
 client/dist/bundle.js
+client/src/firebaseConfig.js
