+++ client/src/firebaseConfig.js
@@ -1,13 +1,15 @@
 const firebase = require('firebase/app');
+
 const config = {
-    apiKey: process.env.APIKEY,
-    authDomain: process.env.DOMAIN,
-    databaseURL: process.env.DBURL,
-    projectId: process.env.PROJECTID,
-    storageBucket: process.env.STORAGEBUCKET,
-    messagingSenderId: process.env.MESSAGINGSENDERID,
-    appId: process.env.APPID
+    apiKey: 'AIzaSyANFsNxSnbvx8JYla2srLVtIS2tIOGSPMA',
+    authDomain: 'found-dog.firebaseapp.com',
+    databaseURL: 'https://found-dog.firebaseio.com',
+    projectId: 'found-dog',
+    storageBucket: 'found-dog.appspot.com',
+    messagingSenderId: '190867854845',
+    appId: ':190867854845:web:c995c8b214fa6fe4'
 };
+
 firebase.initializeApp(config);
 
 const provider = new firebase.auth.GoogleAuthProvider()
@@ -16,5 +18,5 @@ const auth = firebase.auth();
 module.exports = {
     firebase,
     provider,
-    auth
+    auth,
 }
