+++ database-psql/index.js
@@ -1,10 +1,11 @@
 const Sequelize = require('sequelize');
 const bcrypt = require('bcrypt');
 require('dotenv').config();
-
+console.log(process.env.MY_HOST);
 const sequelize = new Sequelize(`${process.env.MY_DB}`, `${process.env.MY_UN}`, `${process.env.MY_PW}`, {
   host: `${process.env.MY_HOST}`,
   dialect: 'postgres',
+  port: process.env.PORT // add port
 });
 
 //require bcrypt
