+++ server/app.js
@@ -4,16 +4,20 @@ const express = require('express');
 const path = require('path');
 //access routes in router file
 const router  = require('../routers/routes.js');
-require('../database-psql/index.js');
+const bodyParser = require('body-parser');
 require('dotenv').config();
+//
+
+require('../database-psql/index.js');
+
 // call express and assign to variable
 const app = express();
 //dotenv
-require('dotenv').config();
 
 app.set('view engine', 'ejs');
 //middleware
 //serve static files
+app.use(bodyParser.json())
 app.use(express.static(path.join(__dirname, '../client/dist')));
 
 //use routes in router file
